move test to ttnn
