+++ tests/tt_metal/CMakeLists.txt
@@ -2,4 +2,3 @@ add_subdirectory(distributed)
 add_subdirectory(tt_fabric)
 add_subdirectory(tt_metal)
 add_subdirectory(multihost)
-add_subdirectory(udm)
