+++ tests/ttnn/unit_tests/gtests/udm/kernels/compute_reduce.cpp
