+++ tests/ttnn/unit_tests/gtests/udm/test_udm_copy.cpp
@@ -5,7 +5,7 @@
 #include <gtest/gtest.h>
 
 #include "tests/tt_metal/tt_metal/common/multi_device_fixture.hpp"
-#include "tests/tt_metal/udm/test_udm_utils.hpp"
+#include "tests/ttnn/unit_tests/gtests/udm/test_udm_utils.hpp"
 
 #include "tt_metal/udm/mesh_kernel.hpp"
 #include "tt_metal/udm/mesh_utils.hpp"
@@ -66,7 +66,7 @@ tt::tt_metal::experimental::udm::MeshProgram create_program(
     tt::tt_metal::experimental::udm::MeshKernelHandle kernel_id = tt::tt_metal::experimental::udm::CreateMeshKernel(
         mesh_builder,
         program,
-        "tests/tt_metal/udm/kernels/copy.cpp",
+        "tests/ttnn/unit_tests/gtests/udm/kernels/copy.cpp",
         gcores_info.gcores,
         tt::tt_metal::DataMovementConfig{
             .processor = tt::tt_metal::DataMovementProcessor::RISCV_0,
