+++ tests/ttnn/unit_tests/gtests/udm/test_udm_reduction_interleaved.cpp
@@ -6,7 +6,7 @@
 #include <cmath>
 
 #include "tests/tt_metal/tt_metal/common/multi_device_fixture.hpp"
-#include "tests/tt_metal/udm/test_udm_utils.hpp"
+#include "tests/ttnn/unit_tests/gtests/udm/test_udm_utils.hpp"
 #include "tt_metal/programming_examples/matmul/matmul_common/bmm_op.hpp"
 #include "tt_metal/api/tt-metalium/bfloat16.hpp"
 
@@ -91,7 +91,7 @@ tt::tt_metal::experimental::udm::MeshProgram create_program(
         tt::tt_metal::experimental::udm::CreateMeshKernel(
             mesh_builder,
             program,
-            "tests/tt_metal/udm/kernels/dataflow_reduce.cpp",
+            "tests/ttnn/unit_tests/gtests/udm/kernels/dataflow_reduce.cpp",
             gcores_info.gcores,
             tt::tt_metal::DataMovementConfig{
                 .processor = tt::tt_metal::DataMovementProcessor::RISCV_0,
@@ -104,7 +104,7 @@ tt::tt_metal::experimental::udm::MeshProgram create_program(
         tt::tt_metal::experimental::udm::CreateMeshKernel(
             mesh_builder,
             program,
-            "tests/tt_metal/udm/kernels/compute_reduce.cpp",
+            "tests/ttnn/unit_tests/gtests/udm/kernels/compute_reduce.cpp",
             gcores_info.gcores,
             tt::tt_metal::ComputeConfig{
                 .fp32_dest_acc_en = true,  // Use FP32 accumulation for better precision
